Assignment02/requirements.txt
Flask==2.0.2
paho-mqtt